Enhance transaction form and data handling
- Added minimum and step attributes to amount input in TransactionForm for better user experience.
- Implemented getTransaction function to fetch transactions by month and year.
- Updated page component to retrieve and display transactions and categories.
- Modified table styles for improved readability.
- Changed amount field in transactionSchema to numeric for handling decimal values.
- Adjusted insertData function to convert amount to string for database compatibility.

+++ src/db/schema.ts
@@ -1,4 +1,4 @@
-import { date, integer, pgTable, text } from "drizzle-orm/pg-core";
+import { date, integer, numeric, pgTable, text } from "drizzle-orm/pg-core";
 
 export const categorySchema = pgTable("Categories", {
     id: integer().primaryKey().generatedAlwaysAsIdentity(),
@@ -11,7 +11,7 @@ export const categorySchema = pgTable("Categories", {
 export const transactionSchema = pgTable("Transactions", {
     id: integer().primaryKey().generatedAlwaysAsIdentity(),
     userId: text("User_Id").notNull(),
-    amount: integer().notNull(),
+    amount: numeric().notNull(), // numeric is being used to allow for the floating point values. Previous integer().notNull() will only allow whole numbers.
     description: text().notNull(),
     transactionDate: date("Transaction_Date").notNull(),
     categoryId: integer("Category_Id").references(() => categorySchema.id).notNull()
